backend_worker/requirements.txt
boto3
PyMuPDF
python-docx
sentence-transformers
scikit-learn
psycopg2-binary